Move TRD track reader/writer to TRDWorkflowIO
