+++ Detectors/TRD/workflow/io/CMakeLists.txt
@@ -15,4 +15,12 @@ o2_add_library(TRDWorkflowIO
                        src/TRDTrackletWriterSpec.cxx
                        src/TRDTrapRawWriterSpec.cxx
                        src/TRDCalibratedTrackletWriterSpec.cxx
-               PUBLIC_LINK_LIBRARIES O2::DataFormatsTRD O2::SimulationDataFormat O2::DPLUtils)
+                       src/TRDTrackWriterSpec.cxx
+                       src/TRDTrackReaderSpec.cxx
+               PUBLIC_LINK_LIBRARIES O2::DataFormatsTRD O2::SimulationDataFormat O2::DPLUtils O2::GPUDataTypeHeaders O2::DataFormatsTPC)
+
+
+o2_add_executable(track-reader
+                 COMPONENT_NAME trd
+                 SOURCES src/trd-track-reader-workflow.cxx
+                 PUBLIC_LINK_LIBRARIES O2::TRDWorkflowIO)
